pyproject.toml
[project]
name = "urban-heat-island"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "folium>=0.18.0",
    "geemap>=0.35.0",
    "geopy>=2.4.1",
    "ipykernel>=6.29.5",
    "ipyleaflet>=0.19.2",
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.2.5",
    "matplotlib>=3.9.2",
    "notebook>=7.2.2",
    "pandas>=2.2.3",
    "pre-commit>=4.0.1",
    "python-decouple>=3.8",
    "setuptools>=75.2.0",
]

[tool.uv]
dev-dependencies = []

[tool.ruff]
line-length = 120
target-version = 'py312'
show-fixes = true

exclude = [
    ".git",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "site-packages",
    "venv",
]
[tool.ruff.format]
quote-style = "double"
docstring-code-format = true
skip-magic-trailing-comma = false

[tool.ruff.lint]
ignore =["D205","D401","D211","D212", "F401"]
